api_migration/requirements.txt
python-pptx
python-docx
ibm-generative-ai
streamlit-code-editor